Added 3 Components, Added Armour Models
Those are the major changes. The components are:
- Obscurity - Gives invisibility, and causes fog to obscure your vision
- Baking - Creates a cake when right clicking the ground.
- Music Player - Plays a music disc track when right clicked. It should
choose it's record from any vanilla record, and also any modded records
that are installed.

Also, Saturation on items was changed to only restore 1 drumstick (2
hunger points) per right click. The armour one will restore all hunger,
but only when it goes down halfway.

I added 5 sets of armour model textures, and changed the API to allow
for more custom armour textures.

I made the golem ID load as a unique ID with the method
findGlobalUniqueEntityId(). That ID loads into the config file, and can
still be changed there. Hopefully that will solve some of the entity ID
problems though.

Included a packaged Artifacts-1.0.4.jar file. I've tested it, and all
seems in order It's working in both 1.7.10 and 1.7.2.

I believe those are all the major changes. Enjoy the new armour
textures. =)